// src/app/Model/Event_Sample.ts
export class EventSample {
  tgbatdau: Date = new Date();
  tgketthuc: string = '';
  noidung: string = '';
  ghichu: string = '';
  chutri: string = '';
  laplai: string = '';
  thanhphan: string = '';
  chuanbi: string = '';
  khachmoi: string = '';
  diadiem: string = '';
  hopkhan: string = '0';
  dsLienQuan: Array<any> = [];
}

export class EventSampleUpdate {
  tgbatdau: Date = new Date();
  tgketthuc: string = '';
  noidung: string = '';
  ghichu: string = '';
  chutri: string = '';
  laplai: string = '';
  thanhphan: string = '';
  chuanbi: string = '';
  khachmoi: string = '';
  diadiem: string = '';
  hopkhan: string = '0';
  dsLienQuan: Array<any> = [];
  idLichTuanMau: string = '';
}